Refactor unit tests by removing unused mock fixtures and improving test clarity. Updated tests across various modules including account balance, B2B, B2C, and bill manager to streamline dependencies and enhance maintainability.
