+++ tests/unit/auth/test_token_manager.py
@@ -5,11 +5,9 @@ These tests cover token retrieval, caching, and error handling for both synchron
 
 import pytest
 from datetime import datetime, timedelta
-from unittest.mock import MagicMock , AsyncMock , patch
-from mpesakit.http_client import HttpClient,AsyncHttpClient
+from unittest.mock import patch
 from mpesakit.auth import TokenManager , AsyncTokenManager
 from mpesakit.errors import MpesaApiException, MpesaError
-
 
 @pytest.fixture
 def valid_credentials():
@@ -19,7 +17,6 @@ def valid_credentials():
         "consumer_secret": "test_secret",
     }
 
-
 @pytest.fixture
 def invalid_credentials():
     """Provide invalid M-Pesa credentials for testing."""
@@ -27,20 +24,6 @@ def invalid_credentials():
         "consumer_key": "invalid_key",
         "consumer_secret": "invalid_secret",
     }
-
-
-@pytest.fixture
-def mock_http_client():
-    """Provide a MagicMock HttpClient for testing."""
-    client = MagicMock(spec=HttpClient)
-    return client
-
-@pytest.fixture
-def mock_async_http_client():
-    """Provide a AsyncMock AsyncHttpClient for testing."""
-    client = AsyncMock(spec=AsyncHttpClient)
-    return client
-
 
 def test_get_token_success(valid_credentials, mock_http_client):
     """Test that a valid token can be retrieved."""
@@ -55,7 +38,6 @@ def test_get_token_success(valid_credentials, mock_http_client):
     )
     token = tm.get_token()
     assert token == "mocked_token_1234567890"
-
 
 def test_token_caching(valid_credentials, mock_http_client):
     """Test that the token is cached and reused until it expires."""
@@ -72,7 +54,6 @@ def test_token_caching(valid_credentials, mock_http_client):
     token2 = tm.get_token()
     assert token1 == token2
 
-
 def test_force_refresh_token(valid_credentials, mock_http_client):
     """Test that forcing a token refresh retrieves a new token."""
     mock_http_client.get.side_effect = [
@@ -88,7 +69,6 @@ def test_force_refresh_token(valid_credentials, mock_http_client):
     token2 = tm.get_token(force_refresh=True)
     assert token1 == "token1"
     assert token2 == "token2"
-
 
 def test_invalid_credentials_raises(mock_http_client, invalid_credentials):
     """Test that invalid credentials raise an exception."""
@@ -111,7 +91,6 @@ def test_invalid_credentials_raises(mock_http_client, invalid_credentials):
         or excinfo.value.error.status_code == 403
     )
 
-
 def test_invalid_grant_type(valid_credentials, mock_http_client, monkeypatch):
     """Test that an invalid grant type raises an exception."""
     tm = TokenManager(
@@ -129,7 +108,6 @@ def test_invalid_grant_type(valid_credentials, mock_http_client, monkeypatch):
     with pytest.raises(MpesaApiException) as excinfo:
         tm.get_token(force_refresh=True)
     assert excinfo.value.error.status_code == 403
-
 
 def test_invalid_auth_type(valid_credentials, mock_http_client, monkeypatch):
     """Test that an invalid auth type raises an exception."""
@@ -149,7 +127,6 @@ def test_invalid_auth_type(valid_credentials, mock_http_client, monkeypatch):
     with pytest.raises(MpesaApiException) as excinfo:
         tm.get_token(force_refresh=True)
     assert excinfo.value.error.status_code == 403
-
 
 def test_mpesa_api_exception_with_empty_error_message(
     valid_credentials, mock_http_client, monkeypatch
@@ -177,7 +154,6 @@ def test_mpesa_api_exception_with_empty_error_message(
     assert err.error_code == "AUTH_INVALID_CREDENTIALS"
     assert "Invalid credentials" in err.error_message
     assert err.status_code == 400
-
 
 def test_token_missing_raises_exception(
     valid_credentials, mock_http_client, monkeypatch
@@ -216,7 +192,6 @@ async def test_async_get_token_success(valid_credentials, mock_async_http_client
     assert token == "mocked_async_token"
     mock_async_http_client.get.assert_called_once()
 
-
 @pytest.mark.asyncio
 async def test_async_token_caching(valid_credentials, mock_async_http_client):
     """Test that the token is cached and reused asynchronously."""
@@ -233,7 +208,6 @@ async def test_async_token_caching(valid_credentials, mock_async_http_client):
     token2 = await tm.get_token()
     assert token1 == token2
     mock_async_http_client.get.assert_called_once()
-
 
 @pytest.mark.asyncio
 async def test_async_force_refresh_token(valid_credentials, mock_async_http_client):
@@ -252,7 +226,6 @@ async def test_async_force_refresh_token(valid_credentials, mock_async_http_clie
     assert token1 == "async_token1"
     assert token2 == "async_token2"
     assert mock_async_http_client.get.call_count == 2
-
 
 @pytest.mark.asyncio
 @patch("mpesakit.auth.token_manager.datetime")
@@ -279,7 +252,6 @@ async def test_async_expired_token_refresh(mock_dt, valid_credentials, mock_asyn
     assert token == "refreshed_async_token"
     assert mock_async_http_client.get.call_count == 2
 
-
 @pytest.mark.asyncio
 async def test_async_invalid_credentials_raises_custom_error(valid_credentials, mock_async_http_client):
     """Test the specific async logic for empty 400 response being converted to a detailed MpesaApiException."""
@@ -304,7 +276,6 @@ async def test_async_invalid_credentials_raises_custom_error(valid_credentials, 
     assert "Invalid credentials" in err.error_message
     assert err.status_code == 400
 
-
 @pytest.mark.asyncio
 async def test_async_invalid_grant_type(valid_credentials, mock_async_http_client, monkeypatch):
     """Test that an invalid grant type raises an exception asynchronously."""
@@ -323,7 +294,6 @@ async def test_async_invalid_grant_type(valid_credentials, mock_async_http_clien
     with pytest.raises(MpesaApiException) as excinfo:
         await tm.get_token(force_refresh=True)
     assert excinfo.value.error.status_code == 403
-
 
 @pytest.mark.asyncio
 async def test_async_invalid_auth_type(valid_credentials, mock_async_http_client, monkeypatch):
