+++ tests/unit/b2c_account_top_up/test_b2c_account_top_up.py
@@ -4,11 +4,8 @@ This module tests the B2C Account TopUp API client, ensuring it can initiate top
 process responses correctly, and manage callback/error cases.
 """
 
-from unittest.mock import AsyncMock, MagicMock
-
 import pytest
 
-from mpesakit.auth import AsyncTokenManager, TokenManager
 from mpesakit.b2c_account_top_up import (
     AsyncB2CAccountTopUp,
     B2CAccountTopUp,
@@ -19,22 +16,6 @@ from mpesakit.b2c_account_top_up import (
     B2CAccountTopUpTimeoutCallback,
     B2CAccountTopUpTimeoutCallbackResponse,
 )
-from mpesakit.http_client import AsyncHttpClient, HttpClient
-
-
-@pytest.fixture
-def mock_token_manager():
-    """Mock TokenManager to return a fixed token."""
-    mock = MagicMock(spec=TokenManager)
-    mock.get_token.return_value = "test_token"
-    return mock
-
-
-@pytest.fixture
-def mock_http_client():
-    """Mock HttpClient to simulate HTTP requests."""
-    return MagicMock(spec=HttpClient)
-
 
 @pytest.fixture
 def b2c_account_topup(mock_http_client, mock_token_manager):
@@ -42,7 +23,6 @@ def b2c_account_topup(mock_http_client, mock_token_manager):
     return B2CAccountTopUp(
         http_client=mock_http_client, token_manager=mock_token_manager
     )
-
 
 def valid_b2c_account_topup_request():
     """Create a valid B2CAccountTopUpRequest for testing."""
@@ -58,7 +38,6 @@ def valid_b2c_account_topup_request():
         QueueTimeOutURL="https://mydomain/path/timeout",
         ResultURL="https://mydomain/path/result",
     )
-
 
 def test_topup_success(b2c_account_topup, mock_http_client):
     """Test that topup request is acknowledged and successful."""
@@ -78,7 +57,6 @@ def test_topup_success(b2c_account_topup, mock_http_client):
     assert response.ResponseCode == response_data["ResponseCode"]
     assert response.ResponseDescription == response_data["ResponseDescription"]
 
-
 def test_topup_http_error(b2c_account_topup, mock_http_client):
     """Test handling of HTTP errors during topup request."""
     request = valid_b2c_account_topup_request()
@@ -86,7 +64,6 @@ def test_topup_http_error(b2c_account_topup, mock_http_client):
     with pytest.raises(Exception) as excinfo:
         b2c_account_topup.topup(request)
     assert "HTTP error" in str(excinfo.value)
-
 
 def test_b2c_account_topup_success_callback():
     """Test parsing of a successful B2C Account TopUp callback."""
@@ -117,7 +94,6 @@ def test_b2c_account_topup_success_callback():
     assert callback.Result.TransactionID == "QKA81LK5CY"
     assert callback.Result.ResultDesc.startswith("The service request is processed")
 
-
 def test_b2c_account_topup_fail_callback():
     """Test parsing of a failed B2C Account TopUp callback."""
     payload = {
@@ -135,13 +111,11 @@ def test_b2c_account_topup_fail_callback():
     assert "cancelled" in callback.Result.ResultDesc
     assert callback.Result.TransactionID == "TX123456"
 
-
 def test_b2c_account_topup_callback_response():
     """Test the response schema for B2C Account TopUp callback."""
     resp = B2CAccountTopUpCallbackResponse()
     assert resp.ResultCode == 0
     assert "processed successfully" in resp.ResultDesc
-
 
 def test_b2c_account_topup_timeout_callback():
     """Test parsing of a B2C Account TopUp timeout callback."""
@@ -159,13 +133,11 @@ def test_b2c_account_topup_timeout_callback():
     assert callback.Result.ResultCode == "1"
     assert "timed out" in callback.Result.ResultDesc
 
-
 def test_b2c_account_topup_timeout_callback_response():
     """Test the response schema for B2C Account TopUp timeout callback."""
     resp = B2CAccountTopUpTimeoutCallbackResponse()
     assert resp.ResultCode == 0
     assert "Timeout notification received" in resp.ResultDesc
-
 
 @pytest.mark.parametrize("result_code_str, expected", [("0", True), ("1", False)])
 def test_b2c_account_topup_string_result_code_is_successful(result_code_str, expected):
@@ -184,28 +156,12 @@ def test_b2c_account_topup_string_result_code_is_successful(result_code_str, exp
     # Should not raise a TypeError when comparing string vs int inside is_successful
     assert callback.is_successful() is expected
 
-
-@pytest.fixture
-def mock_async_token_manager():
-    """Mock AsyncTokenManager to return a fixed token."""
-    mock = AsyncMock(spec=AsyncTokenManager)
-    mock.get_token.return_value = "test_async_token"
-    return mock
-
-
-@pytest.fixture
-def mock_async_http_client():
-    """Mock AsyncHttpClient to simulate async HTTP requests."""
-    return AsyncMock(spec=AsyncHttpClient)
-
-
 @pytest.fixture
 def async_b2c_account_topup(mock_async_http_client, mock_async_token_manager):
     """Fixture to create an AsyncB2CAccountTopUp instance with mocked dependencies."""
     return AsyncB2CAccountTopUp(
         http_client=mock_async_http_client, token_manager=mock_async_token_manager
     )
-
 
 @pytest.mark.asyncio
 async def test_async_topup_success(
@@ -228,7 +184,6 @@ async def test_async_topup_success(
     assert response.is_successful() is True
     assert response.ResponseCode == response_data["ResponseCode"]
 
-
 @pytest.mark.asyncio
 async def test_async_topup_http_error(
     async_b2c_account_topup, mock_async_http_client, mock_async_token_manager
@@ -241,7 +196,6 @@ async def test_async_topup_http_error(
     with pytest.raises(Exception) as excinfo:
         await async_b2c_account_topup.topup(request)
     assert "Async HTTP error" in str(excinfo.value)
-
 
 @pytest.mark.asyncio
 async def test_async_topup_token_retrieval(
