+++ tests/unit/bill_manager/test_bill_manager.py
@@ -5,12 +5,10 @@ invoicing, cancellation, and error cases.
 """
 
 from datetime import datetime
-from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 from pydantic import ValidationError
 
-from mpesakit.auth import AsyncTokenManager, TokenManager
 from mpesakit.bill_manager.bill_manager import AsyncBillManager, BillManager
 from mpesakit.bill_manager.schemas import (
     BillManagerBulkInvoiceRequest,
@@ -25,22 +23,6 @@ from mpesakit.bill_manager.schemas import (
     BillManagerUpdateOptInRequest,
     BillManagerUpdateOptInResponse,
 )
-from mpesakit.http_client import AsyncHttpClient, HttpClient
-
-
-@pytest.fixture
-def mock_token_manager():
-    """Mock TokenManager for testing purposes."""
-    mock = MagicMock(spec=TokenManager)
-    mock.get_token.return_value = "test_token"
-    return mock
-
-
-@pytest.fixture
-def mock_http_client():
-    """Mock HttpClient for testing purposes."""
-    return MagicMock(spec=HttpClient)
-
 
 @pytest.fixture
 def bill_manager(mock_http_client, mock_token_manager):
@@ -50,7 +32,6 @@ def bill_manager(mock_http_client, mock_token_manager):
         token_manager=mock_token_manager,
         app_key="test_app_key",
     )
-
 
 def valid_opt_in_request():
     """Creates a valid opt-in request for Bill Manager."""
@@ -63,7 +44,6 @@ def valid_opt_in_request():
         callbackurl="http://my.server.com/bar/callback",
     )
 
-
 def valid_update_opt_in_request():
     """Creates a valid update opt-in request for Bill Manager."""
     return BillManagerUpdateOptInRequest(
@@ -74,7 +54,6 @@ def valid_update_opt_in_request():
         logo="image",
         callbackurl="http://my.server.com/bar/callback",
     )
-
 
 def valid_single_invoice_request():
     """Creates a valid single invoice request for Bill Manager."""
@@ -93,16 +72,13 @@ def valid_single_invoice_request():
         ],
     )
 
-
 def valid_bulk_invoice_request():
     """Creates a valid bulk invoice request for Bill Manager."""
     return BillManagerBulkInvoiceRequest(invoices=[valid_single_invoice_request()])
 
-
 def valid_cancel_single_invoice_request():
     """Creates a valid cancel single invoice request for Bill Manager."""
     return BillManagerCancelSingleInvoiceRequest(externalReference="113")
-
 
 def valid_cancel_bulk_invoice_request():
     """Creates a valid cancel bulk invoice request for Bill Manager."""
@@ -112,7 +88,6 @@ def valid_cancel_bulk_invoice_request():
             BillManagerCancelSingleInvoiceRequest(externalReference="114"),
         ]
     )
-
 
 def test_opt_in_success(bill_manager, mock_http_client):
     """Test successful opt-in to Bill Manager."""
@@ -128,7 +103,6 @@ def test_opt_in_success(bill_manager, mock_http_client):
     assert response.app_key == response_data["app_key"]
     assert response.rescode == "200"
 
-
 def test_update_opt_in_success(bill_manager, mock_http_client):
     """Test successful update of opt-in settings for Bill Manager."""
     request = valid_update_opt_in_request()
@@ -140,7 +114,6 @@ def test_update_opt_in_success(bill_manager, mock_http_client):
     response = bill_manager.update_opt_in(request)
     assert isinstance(response, BillManagerUpdateOptInResponse)
     assert response.rescode == "200"
-
 
 def test_send_single_invoice_success(bill_manager, mock_http_client):
     """Test sending a single invoice via Bill Manager."""
@@ -156,7 +129,6 @@ def test_send_single_invoice_success(bill_manager, mock_http_client):
     assert response.is_successful() is True
     assert response.Status_Message == response_data["Status_Message"]
 
-
 def test_send_bulk_invoice_success(bill_manager, mock_http_client):
     """Test sending multiple invoices via Bill Manager."""
     request = valid_bulk_invoice_request()
@@ -169,7 +141,6 @@ def test_send_bulk_invoice_success(bill_manager, mock_http_client):
     response = bill_manager.send_bulk_invoice(request)
     assert isinstance(response, BillManagerBulkInvoiceResponse)
     assert response.Status_Message == response_data["Status_Message"]
-
 
 def test_cancel_single_invoice_success(bill_manager, mock_http_client):
     """Test cancelling a single invoice via Bill Manager."""
@@ -186,7 +157,6 @@ def test_cancel_single_invoice_success(bill_manager, mock_http_client):
     assert response.is_successful() is True
     assert response.Status_Message == response_data["Status_Message"]
 
-
 def test_cancel_bulk_invoice_success(bill_manager, mock_http_client):
     """Test cancelling multiple invoices via Bill Manager."""
     request = valid_cancel_bulk_invoice_request()
@@ -201,7 +171,6 @@ def test_cancel_bulk_invoice_success(bill_manager, mock_http_client):
     assert isinstance(response, BillManagerCancelInvoiceResponse)
     assert response.Status_Message == response_data["Status_Message"]
 
-
 def test_bill_manager_http_error(bill_manager, mock_http_client):
     """Test handling of HTTP errors when sending a single invoice."""
     request = valid_single_invoice_request()
@@ -209,7 +178,6 @@ def test_bill_manager_http_error(bill_manager, mock_http_client):
     with pytest.raises(Exception) as excinfo:
         bill_manager.send_single_invoice(request)
     assert "HTTP error" in str(excinfo.value)
-
 
 def test_app_key_required_for_invoice(mock_http_client, mock_token_manager):
     """Test app_key requirement for sending a single invoice."""
@@ -220,7 +188,6 @@ def test_app_key_required_for_invoice(mock_http_client, mock_token_manager):
     with pytest.raises(ValueError) as excinfo:
         manager.send_single_invoice(request)
     assert "app_key must be set" in str(excinfo.value)
-
 
 @pytest.mark.parametrize(
     "due_date,expected",
@@ -278,7 +245,6 @@ def test_due_date_valid_formats(due_date, expected):
         microsecond=(dt_result.microsecond // 10000) * 10000
     ) == dt_expected.replace(microsecond=(dt_expected.microsecond // 10000) * 10000)
 
-
 @pytest.mark.parametrize(
     "due_date",
     [
@@ -308,7 +274,6 @@ def test_due_date_invalid_formats_raise(due_date):
         BillManagerSingleInvoiceRequest(**req)
     assert "validation error" in str(excinfo.value)
 
-
 def test_due_date_missing_raises():
     """Test missing dueDate raises ValueError."""
     req = {
@@ -326,7 +291,6 @@ def test_due_date_missing_raises():
         BillManagerSingleInvoiceRequest.model_validate(req)
     assert "dueDate is required" in str(excinfo.value)
 
-
 def test_billed_period_invalid_raises():
     """Test invalid billedPeriod raises ValueError."""
     req = {
@@ -343,7 +307,6 @@ def test_billed_period_invalid_raises():
     with pytest.raises(ValueError) as excinfo:
         BillManagerSingleInvoiceRequest(**req)
     assert "billedPeriod" in str(excinfo.value)
-
 
 def test_result_code_as_string_does_not_raise(bill_manager, mock_http_client):
     """Ensure response.resultCode as a string does not cause type errors in is_successful."""
@@ -364,21 +327,6 @@ def test_result_code_as_string_does_not_raise(bill_manager, mock_http_client):
     is_success = response.is_successful()
     assert isinstance(is_success, bool)
 
-
-@pytest.fixture
-def mock_async_token_manager():
-    """Mock AsyncTokenManager for testing purposes."""
-    mock = AsyncMock(spec=AsyncTokenManager)
-    mock.get_token.return_value = "test_async_token"
-    return mock
-
-
-@pytest.fixture
-def mock_async_http_client():
-    """Mock AsyncHttpClient for testing purposes."""
-    return AsyncMock(spec=AsyncHttpClient)
-
-
 @pytest.fixture
 def async_bill_manager(mock_async_http_client, mock_async_token_manager):
     """Fixture to create an AsyncBillManager instance with mocked AsyncHttpClient and AsyncTokenManager."""
@@ -387,7 +335,6 @@ def async_bill_manager(mock_async_http_client, mock_async_token_manager):
         token_manager=mock_async_token_manager,
         app_key="test_app_key",
     )
-
 
 @pytest.mark.asyncio
 async def test_async_opt_in_success(async_bill_manager, mock_async_http_client):
@@ -404,7 +351,6 @@ async def test_async_opt_in_success(async_bill_manager, mock_async_http_client):
     assert response.app_key == response_data["app_key"]
     assert response.rescode == "200"
 
-
 @pytest.mark.asyncio
 async def test_async_update_opt_in_success(async_bill_manager, mock_async_http_client):
     """Test successful async update of opt-in settings for Bill Manager."""
@@ -417,7 +363,6 @@ async def test_async_update_opt_in_success(async_bill_manager, mock_async_http_c
     response = await async_bill_manager.update_opt_in(request)
     assert isinstance(response, BillManagerUpdateOptInResponse)
     assert response.rescode == "200"
-
 
 @pytest.mark.asyncio
 async def test_async_send_single_invoice_success(
@@ -436,7 +381,6 @@ async def test_async_send_single_invoice_success(
     assert response.is_successful() is True
     assert response.Status_Message == response_data["Status_Message"]
 
-
 @pytest.mark.asyncio
 async def test_async_send_bulk_invoice_success(
     async_bill_manager, mock_async_http_client
@@ -452,7 +396,6 @@ async def test_async_send_bulk_invoice_success(
     response = await async_bill_manager.send_bulk_invoice(request)
     assert isinstance(response, BillManagerBulkInvoiceResponse)
     assert response.Status_Message == response_data["Status_Message"]
-
 
 @pytest.mark.asyncio
 async def test_async_cancel_single_invoice_success(
@@ -472,7 +415,6 @@ async def test_async_cancel_single_invoice_success(
     assert response.is_successful() is True
     assert response.Status_Message == response_data["Status_Message"]
 
-
 @pytest.mark.asyncio
 async def test_async_cancel_bulk_invoice_success(
     async_bill_manager, mock_async_http_client
@@ -490,7 +432,6 @@ async def test_async_cancel_bulk_invoice_success(
     assert isinstance(response, BillManagerCancelInvoiceResponse)
     assert response.Status_Message == response_data["Status_Message"]
 
-
 @pytest.mark.asyncio
 async def test_async_bill_manager_http_error(
     async_bill_manager, mock_async_http_client
@@ -501,7 +442,6 @@ async def test_async_bill_manager_http_error(
     with pytest.raises(Exception) as excinfo:
         await async_bill_manager.send_single_invoice(request)
     assert "HTTP error" in str(excinfo.value)
-
 
 @pytest.mark.asyncio
 async def test_async_app_key_required_for_invoice(
