+++ tests/unit/business_buy_goods/test_business_buy_goods.py
@@ -4,11 +4,8 @@ This module tests the Business Buy Goods API client, ensuring it can handle paym
 process responses correctly, and manage error cases.
 """
 
-from unittest.mock import AsyncMock, MagicMock
-
 import pytest
 
-from mpesakit.auth import AsyncTokenManager, TokenManager
 from mpesakit.business_buy_goods import (
     AsyncBusinessBuyGoods,
     BusinessBuyGoods,
@@ -19,22 +16,6 @@ from mpesakit.business_buy_goods import (
     BusinessBuyGoodsTimeoutCallback,
     BusinessBuyGoodsTimeoutCallbackResponse,
 )
-from mpesakit.http_client import AsyncHttpClient, HttpClient
-
-
-@pytest.fixture
-def mock_token_manager():
-    """Mock TokenManager to return a fixed token."""
-    mock = MagicMock(spec=TokenManager)
-    mock.get_token.return_value = "test_token"
-    return mock
-
-
-@pytest.fixture
-def mock_http_client():
-    """Mock HttpClient to simulate HTTP requests."""
-    return MagicMock(spec=HttpClient)
-
 
 @pytest.fixture
 def business_buy_goods(mock_http_client, mock_token_manager):
@@ -42,7 +23,6 @@ def business_buy_goods(mock_http_client, mock_token_manager):
     return BusinessBuyGoods(
         http_client=mock_http_client, token_manager=mock_token_manager
     )
-
 
 def valid_business_buy_goods_request():
     """Create a valid BusinessBuyGoodsRequest for testing."""
@@ -57,7 +37,6 @@ def valid_business_buy_goods_request():
         QueueTimeOutURL="https://mydomain.com/b2b/buygoods/queue/",
         ResultURL="https://mydomain.com/b2b/buygoods/result/",
     )
-
 
 def test_buy_goods_request_acknowledged(business_buy_goods, mock_http_client):
     """Test that buy goods request is acknowledged, not finalized."""
@@ -81,7 +60,6 @@ def test_buy_goods_request_acknowledged(business_buy_goods, mock_http_client):
     assert response.ResponseCode == response_data["ResponseCode"]
     assert response.ResponseDescription == response_data["ResponseDescription"]
 
-
 def test_buy_goods_http_error(business_buy_goods, mock_http_client):
     """Test handling of HTTP errors during buy goods request."""
     request = valid_business_buy_goods_request()
@@ -89,7 +67,6 @@ def test_buy_goods_http_error(business_buy_goods, mock_http_client):
     with pytest.raises(Exception) as excinfo:
         business_buy_goods.buy_goods(request)
     assert "HTTP error" in str(excinfo.value)
-
 
 def test_business_buy_goods_result_callback_success():
     """Test parsing of a successful business buy goods result callback."""
@@ -119,13 +96,11 @@ def test_business_buy_goods_result_callback_success():
     assert callback.Result.TransactionID == "QKA81LK5CY"
     assert callback.Result.ResultParameters.ResultParameter[0].Key == "Amount"
 
-
 def test_business_buy_goods_result_callback_response():
     """Test the response schema for result callback."""
     resp = BusinessBuyGoodsResultCallbackResponse()
     assert resp.ResultCode == 0
     assert "Callback received successfully" in resp.ResultDesc
-
 
 def test_business_buy_goods_timeout_callback():
     """Test parsing of a business buy goods timeout callback."""
@@ -143,13 +118,11 @@ def test_business_buy_goods_timeout_callback():
     assert callback.Result.ResultCode == 1
     assert "timed out" in callback.Result.ResultDesc
 
-
 def test_business_buy_goods_timeout_callback_response():
     """Test the response schema for timeout callback."""
     resp = BusinessBuyGoodsTimeoutCallbackResponse()
     assert resp.ResultCode == 0
     assert "Timeout notification received" in resp.ResultDesc
-
 
 def test_business_buy_goods_result_callback_resultcode_string():
     """Ensure a string ResultCode does not raise when checking success."""
@@ -176,28 +149,12 @@ def test_business_buy_goods_result_callback_resultcode_string():
     # Also assert it is considered successful.
     assert callback.is_successful() is True
 
-
-@pytest.fixture
-def mock_async_token_manager():
-    """Mock AsyncTokenManager to return a fixed token."""
-    mock = AsyncMock(spec=AsyncTokenManager)
-    mock.get_token.return_value = "test_token"
-    return mock
-
-
-@pytest.fixture
-def mock_async_http_client():
-    """Mock AsyncHttpClient to simulate async HTTP requests."""
-    return AsyncMock(spec=AsyncHttpClient)
-
-
 @pytest.fixture
 def async_business_buy_goods(mock_async_http_client, mock_async_token_manager):
     """Fixture to create an AsyncBusinessBuyGoods instance with mocked dependencies."""
     return AsyncBusinessBuyGoods(
         http_client=mock_async_http_client, token_manager=mock_async_token_manager
     )
-
 
 @pytest.mark.asyncio
 async def test_async_buy_goods_request_acknowledged(
@@ -219,7 +176,6 @@ async def test_async_buy_goods_request_acknowledged(
     assert response.is_successful() is True
     assert response.ConversationID == response_data["ConversationID"]
 
-
 @pytest.mark.asyncio
 async def test_async_buy_goods_http_error(
     async_business_buy_goods, mock_async_http_client
@@ -232,7 +188,6 @@ async def test_async_buy_goods_http_error(
         await async_business_buy_goods.buy_goods(request)
 
     assert "Async HTTP error" in str(excinfo.value)
-
 
 @pytest.mark.asyncio
 async def test_async_buy_goods_token_manager_called(
@@ -251,7 +206,6 @@ async def test_async_buy_goods_token_manager_called(
     await async_business_buy_goods.buy_goods(request)
 
     mock_async_token_manager.get_token.assert_called_once()
-
 
 @pytest.mark.asyncio
 async def test_async_buy_goods_http_client_called_with_correct_params(
