+++ tests/unit/conftest.py
@@ -1,0 +1,35 @@
+"""Shared fixtures for unit tests."""
+
+import pytest
+from unittest.mock import AsyncMock, MagicMock
+
+from mpesakit.auth import AsyncTokenManager, TokenManager
+from mpesakit.http_client import AsyncHttpClient, HttpClient
+
+
+@pytest.fixture
+def mock_token_manager():
+    """Mock TokenManager to return a fixed token."""
+    mock = MagicMock(spec=TokenManager)
+    mock.get_token.return_value = "test_token"
+    return mock
+
+
+@pytest.fixture
+def mock_http_client():
+    """Mock HttpClient to simulate HTTP requests."""
+    return MagicMock(spec=HttpClient)
+
+
+@pytest.fixture
+def mock_async_token_manager():
+    """Mock AsyncTokenManager to return a fixed token."""
+    mock = AsyncMock(spec=AsyncTokenManager)
+    mock.get_token.return_value = "test_token"
+    return mock
+
+
+@pytest.fixture
+def mock_async_http_client():
+    """Mock AsyncHttpClient to simulate async HTTP requests."""
+    return AsyncMock(spec=AsyncHttpClient)
