+++ tests/unit/mpesa_ratiba/test_mpesa_ratiba.py
@@ -4,12 +4,8 @@ This module tests the Standing Order API client, ensuring it can initiate standi
 process responses correctly, and manage callback/error cases.
 """
 
-from unittest.mock import AsyncMock, MagicMock
-
 import pytest
 
-from mpesakit.auth import AsyncTokenManager, TokenManager
-from mpesakit.http_client import AsyncHttpClient, HttpClient
 from mpesakit.mpesa_ratiba import (
     AsyncMpesaRatiba,
     FrequencyEnum,
@@ -22,26 +18,10 @@ from mpesakit.mpesa_ratiba import (
     TransactionTypeEnum,
 )
 
-
-@pytest.fixture
-def mock_token_manager():
-    """Mock TokenManager to return a fixed token."""
-    mock = MagicMock(spec=TokenManager)
-    mock.get_token.return_value = "test_token"
-    return mock
-
-
-@pytest.fixture
-def mock_http_client():
-    """Mock HttpClient to simulate HTTP requests."""
-    return MagicMock(spec=HttpClient)
-
-
 @pytest.fixture
 def mpesa_ratiba(mock_http_client, mock_token_manager):
     """Fixture to create a MpesaRatiba instance with mocked dependencies."""
     return MpesaRatiba(http_client=mock_http_client, token_manager=mock_token_manager)
-
 
 def valid_standing_order_request():
     """Create a valid StandingOrderRequest for testing."""
@@ -59,7 +39,6 @@ def valid_standing_order_request():
         TransactionDesc="Electric Bike",
         Frequency=FrequencyEnum.DAILY,
     )
-
 
 def test_create_standing_order_success(mpesa_ratiba, mock_http_client):
     """Test that standing order request is acknowledged and successful."""
@@ -91,7 +70,6 @@ def test_create_standing_order_success(mpesa_ratiba, mock_http_client):
         == response_data["ResponseHeader"]["responseDescription"]
     )
 
-
 def test_create_standing_order_http_error(mpesa_ratiba, mock_http_client):
     """Test handling of HTTP errors during standing order request."""
     request = valid_standing_order_request()
@@ -99,7 +77,6 @@ def test_create_standing_order_http_error(mpesa_ratiba, mock_http_client):
     with pytest.raises(Exception) as excinfo:
         mpesa_ratiba.create_standing_order(request)
     assert "HTTP error" in str(excinfo.value)
-
 
 def test_standing_order_success_callback():
     """Test parsing of a successful Standing Order callback."""
@@ -131,7 +108,6 @@ def test_standing_order_success_callback():
         for item in callback.ResponseBody.ResponseData
     )
 
-
 def test_standing_order_fail_callback():
     """Test parsing of a failed Standing Order callback."""
     payload = {
@@ -161,13 +137,11 @@ def test_standing_order_fail_callback():
         for item in callback.ResponseBody.ResponseData
     )
 
-
 def test_standing_order_callback_response():
     """Test the response schema for Standing Order callback."""
     resp = StandingOrderCallbackResponse()
     assert resp.ResultCode == "0"
     assert "processed successfully" in resp.ResultDesc
-
 
 def test_standing_order_request_invalid_date_format():
     """Test StandingOrderRequest raises ValueError for invalid date format."""
@@ -188,7 +162,6 @@ def test_standing_order_request_invalid_date_format():
         )
     assert "Date must be in 'yyyymmdd' format" in str(excinfo.value)
 
-
 def test_standing_order_request_invalid_date_value():
     """Test StandingOrderRequest raises ValueError for invalid date value."""
     with pytest.raises(ValueError) as excinfo:
@@ -208,7 +181,6 @@ def test_standing_order_request_invalid_date_value():
         )
     assert "Date must be in 'yyyymmdd' format" in str(excinfo.value)
 
-
 def test_standing_order_request_other_date_value():
     """Test StandingOrderRequest raises ValueError for invalid date value."""
     request = StandingOrderRequest(
@@ -226,7 +198,6 @@ def test_standing_order_request_other_date_value():
         Frequency=FrequencyEnum.MONTHLY,
     )
     assert request.StartDate == "20241205"  # Should normalize to yyyymmdd format
-
 
 def test_invalid_phone_number():
     """Test that invalid phone numbers raise ValueError."""
@@ -246,7 +217,6 @@ def test_invalid_phone_number():
             Frequency=FrequencyEnum.MONTHLY,
         )
     assert "Invalid PartyA phone number" in str(excinfo.value)
-
 
 def test_callback_resultcode_as_string_handled_gracefully():
     """Ensure StandingOrderCallback.is_successful() handles responseCode as a string without TypeError."""
@@ -275,28 +245,12 @@ def test_callback_resultcode_as_string_handled_gracefully():
         )
     assert result is True
 
-
-@pytest.fixture
-def mock_async_token_manager():
-    """Mock AsyncTokenManager to return a fixed token."""
-    mock = AsyncMock(spec=AsyncTokenManager)
-    mock.get_token.return_value = "test_token"
-    return mock
-
-
-@pytest.fixture
-def mock_async_http_client():
-    """Mock AsyncHttpClient to simulate async HTTP requests."""
-    return AsyncMock(spec=AsyncHttpClient)
-
-
 @pytest.fixture
 def async_mpesa_ratiba(mock_async_http_client, mock_async_token_manager):
     """Fixture to create an AsyncMpesaRatiba instance with mocked dependencies."""
     return AsyncMpesaRatiba(
         http_client=mock_async_http_client, token_manager=mock_async_token_manager
     )
-
 
 @pytest.mark.asyncio
 async def test_create_standing_order_success_async(
@@ -330,7 +284,6 @@ async def test_create_standing_order_success_async(
         response.ResponseHeader.responseDescription
         == response_data["ResponseHeader"]["responseDescription"]
     )
-
 
 @pytest.mark.asyncio
 async def test_create_standing_order_http_error_async(
