+++ tests/unit/services/test_b2c_service.py
@@ -1,44 +1,12 @@
 """Unit tests for B2CService class."""
 
 import pytest
-from unittest.mock import AsyncMock, MagicMock
 from mpesakit.services.b2c import B2CService, AsyncB2CService
 from mpesakit.b2c import B2CResponse, B2CCommandIDType
-from mpesakit.auth import TokenManager, AsyncTokenManager
-from mpesakit.http_client import HttpClient, AsyncHttpClient
 
 from mpesakit.b2c_account_top_up import (
     B2CAccountTopUpResponse,
 )
-
-
-@pytest.fixture
-def mock_token_manager():
-    """Mock TokenManager to return a fixed token."""
-    mock = MagicMock(spec=TokenManager)
-    mock.get_token.return_value = "test_token"
-    return mock
-
-
-@pytest.fixture
-def mock_http_client():
-    """Mock HttpClient to simulate HTTP requests."""
-    return MagicMock(spec=HttpClient)
-
-
-@pytest.fixture
-def mock_async_token_manager():
-    """Mock AsyncTokenManager to return a fixed token."""
-    mock = AsyncMock(spec=AsyncTokenManager)
-    mock.get_token.return_value = "async_test_token"
-    return mock
-
-
-@pytest.fixture
-def mock_async_http_client():
-    """Mock AsyncHttpClient to simulate async HTTP requests."""
-    return AsyncMock(spec=AsyncHttpClient)
-
 
 @pytest.fixture
 def b2c_service(mock_http_client, mock_token_manager):
@@ -48,7 +16,6 @@ def b2c_service(mock_http_client, mock_token_manager):
         token_manager=mock_token_manager,
     )
 
-
 @pytest.fixture
 def async_b2c_service(mock_async_http_client, mock_async_token_manager):
     """Fixture to create an AsyncB2CService instance with mocked dependencies."""
@@ -56,7 +23,6 @@ def async_b2c_service(mock_async_http_client, mock_async_token_manager):
         http_client=mock_async_http_client,
         token_manager=mock_async_token_manager,
     )
-
 
 def test_send_payment_calls_b2c_send_payment(b2c_service, mock_http_client):
     """Test that send_payment calls the B2C service."""
@@ -85,7 +51,6 @@ def test_send_payment_calls_b2c_send_payment(b2c_service, mock_http_client):
     assert resp.ResponseCode == "0"
     assert resp.ResponseDescription == "Request accepted successfully."
 
-
 def test_account_topup_calls_account_topup(b2c_service, mock_http_client):
     """Test that account_topup calls the B2CAccountTopUp service."""
     response_data = {
@@ -111,7 +76,6 @@ def test_account_topup_calls_account_topup(b2c_service, mock_http_client):
     assert isinstance(resp, B2CAccountTopUpResponse)
     assert resp.ResponseCode == "0"
     assert resp.ResponseDescription == "TopUp accepted successfully."
-
 
 def test_send_payment_filters_kwargs(b2c_service, mock_http_client):
     """Test that send_payment filters out unexpected kwargs."""
@@ -140,7 +104,6 @@ def test_send_payment_filters_kwargs(b2c_service, mock_http_client):
     assert resp.ResponseCode == "0"
     assert resp.ResponseDescription == "Request accepted successfully."
 
-
 def test_account_topup_filters_kwargs(b2c_service, mock_http_client):
     """Test that account_topup filters out unexpected kwargs."""
     response_data = {
@@ -167,7 +130,6 @@ def test_account_topup_filters_kwargs(b2c_service, mock_http_client):
     assert resp.ResponseCode == "0"
     assert resp.ResponseDescription == "TopUp accepted successfully."
 
-
 def test_b2c_service_initializes_b2c_correctly(mock_http_client, mock_token_manager):
     """Test B2CService initializes with correct arguments."""
     service = B2CService(
@@ -180,7 +142,6 @@ def test_b2c_service_initializes_b2c_correctly(mock_http_client, mock_token_mana
     if hasattr(service, "b2c"):
         assert service.b2c.http_client is mock_http_client
         assert service.b2c.token_manager is mock_token_manager
-
 
 @pytest.mark.asyncio
 async def test_async_send_payment_calls_b2c_send_payment(
@@ -212,7 +173,6 @@ async def test_async_send_payment_calls_b2c_send_payment(
     assert resp.ResponseCode == "0"
     assert resp.ResponseDescription == "Request accepted successfully."
 
-
 @pytest.mark.asyncio
 async def test_async_account_topup_calls_account_topup(
     async_b2c_service, mock_async_http_client
@@ -241,7 +201,6 @@ async def test_async_account_topup_calls_account_topup(
     assert isinstance(resp, B2CAccountTopUpResponse)
     assert resp.ResponseCode == "0"
     assert resp.ResponseDescription == "TopUp accepted successfully."
-
 
 @pytest.mark.asyncio
 async def test_async_send_payment_filters_kwargs(
@@ -274,7 +233,6 @@ async def test_async_send_payment_filters_kwargs(
     assert resp.ResponseCode == "0"
     assert resp.ResponseDescription == "Request accepted successfully."
 
-
 @pytest.mark.asyncio
 async def test_async_account_topup_filters_kwargs(
     async_b2c_service, mock_async_http_client
@@ -304,7 +262,6 @@ async def test_async_account_topup_filters_kwargs(
     assert isinstance(resp, B2CAccountTopUpResponse)
     assert resp.ResponseCode == "0"
     assert resp.ResponseDescription == "TopUp accepted successfully."
-
 
 def test_async_b2c_service_initializes_b2c_correctly(
     mock_async_http_client, mock_async_token_manager
