+++ tests/unit/services/test_transaction_service.py
@@ -1,26 +1,8 @@
 """Unit tests for TransactionService class."""
 
 import pytest
-from unittest.mock import MagicMock
 from mpesakit.services.transaction import TransactionService
 from mpesakit.transaction_status import TransactionStatusResponse
-from mpesakit.auth import TokenManager
-from mpesakit.http_client import HttpClient
-
-
-@pytest.fixture
-def mock_token_manager():
-    """Mock TokenManager to return a fixed token."""
-    mock = MagicMock(spec=TokenManager)
-    mock.get_token.return_value = "test_token"
-    return mock
-
-
-@pytest.fixture
-def mock_http_client():
-    """Mock HttpClient to simulate HTTP requests."""
-    return MagicMock(spec=HttpClient)
-
 
 @pytest.fixture
 def transaction_service(mock_http_client, mock_token_manager):
@@ -29,7 +11,6 @@ def transaction_service(mock_http_client, mock_token_manager):
         http_client=mock_http_client,
         token_manager=mock_token_manager,
     )
-
 
 def test_query_status_calls_transaction_status(transaction_service, mock_http_client):
     """Test that query_status calls TransactionStatus.query with correct request."""
@@ -95,7 +76,6 @@ def test_query_status_default_remarks(transaction_service, mock_http_client):
     }
     mock_http_client.post.return_value = response_data
 
-
     resp = transaction_service.query_status(
         initiator="testapi",
         security_credential="encrypted_credential",
@@ -113,7 +93,6 @@ def test_query_status_default_remarks(transaction_service, mock_http_client):
     assert isinstance(resp, TransactionStatusResponse)
     assert resp.is_successful() is True
     assert resp.ResponseDescription == "Accept the service request successfully."
-
 
 def test_query_status_filters_kwargs(transaction_service, mock_http_client):
     """Test that query_status filters out unexpected kwargs."""
@@ -142,7 +121,6 @@ def test_query_status_filters_kwargs(transaction_service, mock_http_client):
     assert resp.is_successful() is True
     # unexpected_field should not be present in the response
     assert not hasattr(resp, "unexpected_field")
-
 
 def test_transaction_service_initializes_correctly(
     mock_http_client, mock_token_manager
