+++ tests/unit/transaction_status/test_transaction_status.py
@@ -3,12 +3,8 @@
 This module tests the TransactionStatus class and its methods for querying transaction status.
 """
 
-from unittest.mock import AsyncMock, MagicMock
-
 import pytest
 
-from mpesakit.auth import AsyncTokenManager, TokenManager
-from mpesakit.http_client import AsyncHttpClient, HttpClient
 from mpesakit.transaction_status import (
     AsyncTransactionStatus,
     TransactionStatus,
@@ -20,28 +16,12 @@ from mpesakit.transaction_status import (
     TransactionStatusResultParameter,
 )
 
-
-@pytest.fixture
-def mock_token_manager():
-    """Mock TokenManager for testing."""
-    mock = MagicMock(spec=TokenManager)
-    mock.get_token.return_value = "test_token"
-    return mock
-
-
-@pytest.fixture
-def mock_http_client():
-    """Mock HttpClient for testing."""
-    return MagicMock(spec=HttpClient)
-
-
 @pytest.fixture
 def transaction_status(mock_http_client, mock_token_manager):
     """Fixture to create a TransactionStatus instance with mocked dependencies."""
     return TransactionStatus(
         http_client=mock_http_client, token_manager=mock_token_manager
     )
-
 
 def valid_transaction_status_request():
     """Create a valid TransactionStatusRequest for testing."""
@@ -58,7 +38,6 @@ def valid_transaction_status_request():
         Occasion="JuneSalary",
         OriginalConversationID=None,
     )
-
 
 def test_query_success(transaction_status, mock_http_client):
     """Test querying transaction status with a valid request."""
@@ -85,7 +64,6 @@ def test_query_success(transaction_status, mock_http_client):
     assert args[0] == "/mpesa/transactionstatus/v1/query"
     assert kwargs["headers"]["Authorization"] == "Bearer test_token"
 
-
 def test_query_http_error(transaction_status, mock_http_client):
     """Test handling HTTP errors during transaction status query."""
     request = valid_transaction_status_request()
@@ -93,7 +71,6 @@ def test_query_http_error(transaction_status, mock_http_client):
     with pytest.raises(Exception) as excinfo:
         transaction_status.query(request)
     assert "HTTP error" in str(excinfo.value)
-
 
 def test_transaction_status_response_is_successful_zero_code():
     """Test is_successful method with a zero response code."""
@@ -105,7 +82,6 @@ def test_transaction_status_response_is_successful_zero_code():
     )
     assert resp.is_successful() is True
 
-
 def test_transaction_status_response_is_successful_all_zeros():
     """Test is_successful method with all zeros response code."""
     resp = TransactionStatusResponse(
@@ -115,7 +91,6 @@ def test_transaction_status_response_is_successful_all_zeros():
         ResponseDescription="Accept the service request successfully.",
     )
     assert resp.is_successful() is True
-
 
 def test_transaction_status_response_is_successful_non_zero_code():
     """Test is_successful method with a non-zero response code."""
@@ -127,7 +102,6 @@ def test_transaction_status_response_is_successful_non_zero_code():
     )
     assert resp.is_successful() is False
 
-
 def test_transaction_status_response_is_successful_mixed_code():
     """Test is_successful method with a mixed response code."""
     resp = TransactionStatusResponse(
@@ -138,7 +112,6 @@ def test_transaction_status_response_is_successful_mixed_code():
     )
     assert resp.is_successful() is False
 
-
 def test_transaction_status_response_is_successful_empty_code():
     """Test is_successful method with an empty response code."""
     resp = TransactionStatusResponse(
@@ -148,7 +121,6 @@ def test_transaction_status_response_is_successful_empty_code():
         ResponseDescription="Failed.",
     )
     assert resp.is_successful() is False
-
 
 @pytest.mark.parametrize("invalid_identifier_type", [0, 3, 5, "invalid"])
 def test_transaction_status_request_invalid_identifier_type_raises(
@@ -169,7 +141,6 @@ def test_transaction_status_request_invalid_identifier_type_raises(
     with pytest.raises(ValueError) as excinfo:
         TransactionStatusRequest(**kwargs)
     assert "IdentifierType must be one of" in str(excinfo.value)
-
 
 def test_transaction_status_request_missing_transaction_id_and_originator_conversation_id_raises():
     """Test that missing both TransactionID and OriginalConversationID raises ValueError."""
@@ -192,7 +163,6 @@ def test_transaction_status_request_missing_transaction_id_and_originator_conver
         in str(excinfo.value)
     )
 
-
 def test_transaction_status_request_remarks_too_long_raises():
     """Test that Remarks longer than 100 characters raises ValueError."""
     kwargs = dict(
@@ -209,7 +179,6 @@ def test_transaction_status_request_remarks_too_long_raises():
     with pytest.raises(ValueError) as excinfo:
         TransactionStatusRequest(**kwargs)
     assert "Remarks must not exceed 100 characters." in str(excinfo.value)
-
 
 def test_transaction_status_request_occasion_too_long_raises():
     """Test that Occasion longer than 100 characters raises ValueError."""
@@ -228,7 +197,6 @@ def test_transaction_status_request_occasion_too_long_raises():
     with pytest.raises(ValueError) as excinfo:
         TransactionStatusRequest(**kwargs)
     assert "Occasion must not exceed 100 characters." in str(excinfo.value)
-
 
 def test_transaction_status_request_msisdn_normalization(monkeypatch):
     """Test that PartyA is normalized to a valid Kenyan MSISDN."""
@@ -249,7 +217,6 @@ def test_transaction_status_request_msisdn_normalization(monkeypatch):
     )
     assert req.PartyA == 254712345678
 
-
 def test_transaction_status_request_invalid_msisdn(monkeypatch):
     """Test that invalid PartyA raises ValueError."""
     monkeypatch.setattr("mpesakit.utils.phone.normalize_phone_number", lambda x: None)
@@ -268,11 +235,9 @@ def test_transaction_status_request_invalid_msisdn(monkeypatch):
         TransactionStatusRequest(**kwargs)
     assert "PartyA must be a valid Kenyan MSISDN" in str(excinfo.value)
 
-
 def make_result_parameters(params):
     """Helper function to create ResultParameters from a dictionary."""
     return [TransactionStatusResultParameter(Key=k, Value=v) for k, v in params.items()]
-
 
 def test_result_metadata_properties_all_present():
     """Test that all properties are correctly extracted from ResultParameters."""
@@ -296,7 +261,6 @@ def test_result_metadata_properties_all_present():
     assert meta.transaction_status == "Completed"
     assert meta.transaction_reason == "None"
 
-
 def test_result_metadata_properties_some_missing():
     """Test that properties are correctly extracted when some parameters are missing."""
     params = {
@@ -317,7 +281,6 @@ def test_result_metadata_properties_some_missing():
     assert meta.transaction_status is None
     assert meta.transaction_reason is None
 
-
 def test_result_metadata_properties_none_parameters():
     """Test that properties are None when ResultParameters is empty."""
     meta = TransactionStatusResultMetadata(
@@ -333,7 +296,6 @@ def test_result_metadata_properties_none_parameters():
     assert meta.transaction_receipt is None
     assert meta.transaction_status is None
     assert meta.transaction_reason is None
-
 
 def test_result_callback_schema():
     """Test that TransactionStatusResultCallback schema is correctly defined."""
@@ -356,7 +318,6 @@ def test_result_callback_schema():
     assert callback.Result.transaction_amount == 1000
     assert callback.Result.transaction_receipt == "LKXXXX1234"
     assert callback.Result.transaction_status == "Completed"
-
 
 def test_query_response_code_type_variations(transaction_status, mock_http_client):
     """Ensure TransactionStatusResponse.is_successful handles ResponseCode as str or int without TypeError."""
@@ -383,7 +344,6 @@ def test_query_response_code_type_variations(transaction_status, mock_http_clien
         assert isinstance(resp, TransactionStatusResponse)
         assert resp.is_successful() is expected_success
 
-
 def test_transaction_status_result_callback_is_successful_zero_code():
     """Test is_successful method with ResultCode as '0'."""
     result = TransactionStatusResultMetadata(
@@ -397,7 +357,6 @@ def test_transaction_status_result_callback_is_successful_zero_code():
     )
     callback = TransactionStatusResultCallback(Result=result)
     assert callback.is_successful() is True
-
 
 def test_transaction_status_result_callback_is_successful_all_zeros():
     """Test is_successful method with ResultCode as '00000000'."""
@@ -413,7 +372,6 @@ def test_transaction_status_result_callback_is_successful_all_zeros():
     callback = TransactionStatusResultCallback(Result=result)
     assert callback.is_successful() is True
 
-
 def test_transaction_status_result_callback_is_successful_non_zero_code():
     """Test is_successful method with ResultCode as '1'."""
     result = TransactionStatusResultMetadata(
@@ -427,7 +385,6 @@ def test_transaction_status_result_callback_is_successful_non_zero_code():
     )
     callback = TransactionStatusResultCallback(Result=result)
     assert callback.is_successful() is False
-
 
 def test_transaction_status_result_callback_is_successful_mixed_code():
     """Test is_successful method with ResultCode as '00001'."""
@@ -443,7 +400,6 @@ def test_transaction_status_result_callback_is_successful_mixed_code():
     callback = TransactionStatusResultCallback(Result=result)
     assert callback.is_successful() is False
 
-
 def test_transaction_status_result_callback_is_successful_empty_code():
     """Test is_successful method with an empty ResultCode."""
     result = TransactionStatusResultMetadata(
@@ -457,21 +413,6 @@ def test_transaction_status_result_callback_is_successful_empty_code():
     )
     callback = TransactionStatusResultCallback(Result=result)
     assert callback.is_successful() is False
-
-
-@pytest.fixture
-def mock_async_token_manager():
-    """Mock AsyncTokenManager to return a fixed token."""
-    mock = AsyncMock(spec=AsyncTokenManager)
-    mock.get_token.return_value = "test_token"
-    return mock
-
-
-@pytest.fixture
-def mock_async_http_client():
-    """Mock AsyncHttpClient to simulate async HTTP requests."""
-    return AsyncMock(spec=AsyncHttpClient)
-
 
 @pytest.mark.asyncio
 async def test_async_query_success(mock_async_http_client, mock_async_token_manager):
@@ -503,7 +444,6 @@ async def test_async_query_success(mock_async_http_client, mock_async_token_mana
     args, kwargs = mock_async_http_client.post.call_args
     assert args[0] == "/mpesa/transactionstatus/v1/query"
     assert kwargs["headers"]["Authorization"] == "Bearer test_token"
-
 
 @pytest.mark.asyncio
 async def test_async_query_http_error(mock_async_http_client, mock_async_token_manager):
